// RTOManagement/src/app/models/UserModel.ts
export class UserModel {
  appId: string;
  vehicleType: string;
  userName: string;
  phNo: string;
  email: string;
  dob: string;
  gender: string;
  bloodGroup: string;
  country: string;
  address: string;
  state: string;
  pin: string;
  llNo: string;
  dlNo: string;
  testDriveDate: string;
  licStatus: string;
  applyDate: string;
  expiryDate: string;
  adhar: string;
  bloodGroupCertificate: string;
  certificate: string;
  constructor() {}
}